+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+* Removed the non-functional "map theme" dropdown from appearance settings.
+
 ### Fixed
 
 ## [0.10.1] - 2026-09-04